ml/algos/ARTLab/requirements.txt
# PYTHON 3.10.0
numpy==1.23.5
matplotlib==3.7.1
scikit-learn==1.1.3
torch==2.0.1
torchvision==0.15.2
adversarial-robustness-toolbox==1.14.0
pandas
pillow
kagglehub
